replace old code by the new one and change function style
