+++ web/ui/react-app/src/pages/targets/Targets.tsx
@@ -1,11 +1,11 @@
 import React, { FC } from 'react';
-import { ScrapePoolList2 } from './ScrapePoolList2';
+import ScrapePoolList from './ScrapePoolList';
 
 const Targets: FC = () => {
   return (
     <>
       <h2>Targets</h2>
-      <ScrapePoolList2 />
+      <ScrapePoolList />
     </>
   );
 };
